+++ internal/storage/fake_storage_util.go
@@ -16,6 +16,7 @@ package storage
 
 import (
 	"github.com/fsouza/fake-gcs-server/fakestorage"
+	"github.com/googlecloudplatform/gcsfuse/internal/storage/object"
 )
 
 const port uint16 = 8081
@@ -121,7 +122,7 @@ func getTestFakeStorageObject() []fakestorage.Object {
 			Name:            TestGzipObjectName,
 			Generation:      TestGzipObjectGeneration,
 			Metadata:        map[string]string{MetaDataKey: MetaDataValue},
-			ContentEncoding: ContentEncodingGzip,
+			ContentEncoding: object.ContentEncodingGzip,
 		},
 		Content: []byte(ContentInTestGzipObjectCompressed),
 	}
